Extract regression spec for twitter status xml parsing
